feat(step3): 로또 번호 text 생성자 제거
